pyproject.toml
[project]
name = "innq-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "python-dotenv>=1.0.1",
    "redis",
    "supabase>=2.13.0",
]
